+++ src/transaction/spanner-outbox/runner.ts
@@ -2,36 +2,26 @@ import {
   OutboxTransactionRunner,
   type OutboxEvent,
   type OutboxEventTransaction,
+  type ReadWriteTransactionOptions,
+  type TransactionFn,
 } from '@causa/runtime';
 import { Logger } from '@causa/runtime/nestjs';
 import type { Type } from '@nestjs/common';
 import { SpannerEntityManager } from '../../spanner/index.js';
-import { SpannerStateTransaction } from '../spanner-state-transaction.js';
-import { SpannerTransaction } from '../spanner-transaction.js';
-import { throwRetryableInTransactionIfNeeded } from '../spanner-utils.js';
+import { SpannerReadOnlyStateTransaction } from './readonly-transaction.js';
 import { SpannerOutboxSender } from './sender.js';
+import { throwRetryableInTransactionIfNeeded } from './spanner-utils.js';
+import { SpannerStateTransaction } from './state-transaction.js';
+import { SpannerOutboxTransaction } from './transaction.js';
 
 /**
- * A {@link SpannerTransaction} that uses an {@link OutboxEventTransaction}.
- */
-export type SpannerOutboxTransaction =
-  SpannerTransaction<OutboxEventTransaction>;
-
-/**
- * Option for a function that accepts a {@link SpannerOutboxTransaction}.
- */
-export type SpannerOutboxTransactionOption = {
-  /**
-   * The transaction to use.
-   */
-  readonly transaction?: SpannerOutboxTransaction;
-};
-
-/**
- * An {@link OutboxTransactionRunner} that uses a {@link SpannerTransaction} to run transactions.
+ * An {@link OutboxTransactionRunner} that uses a {@link SpannerOutboxTransaction} to run transactions.
  * Events are stored in a Spanner table before being published.
  */
-export class SpannerOutboxTransactionRunner extends OutboxTransactionRunner<SpannerOutboxTransaction> {
+export class SpannerOutboxTransactionRunner extends OutboxTransactionRunner<
+  SpannerOutboxTransaction,
+  SpannerReadOnlyStateTransaction
+> {
   constructor(
     readonly entityManager: SpannerEntityManager,
     outboxEventType: Type<OutboxEvent>,
@@ -41,9 +31,23 @@ export class SpannerOutboxTransactionRunner extends OutboxTransactionRunner<Span
     super(outboxEventType, sender, logger);
   }
 
+  protected async runReadOnly<RT>(
+    runFn: TransactionFn<SpannerReadOnlyStateTransaction, RT>,
+  ): Promise<RT> {
+    return await this.entityManager.snapshot(async (dbTransaction) => {
+      const transaction = new SpannerReadOnlyStateTransaction(
+        this.entityManager,
+        dbTransaction,
+      );
+
+      return await runFn(transaction);
+    });
+  }
+
   protected async runStateTransaction<RT>(
     eventTransactionFactory: () => OutboxEventTransaction,
-    runFn: (transaction: SpannerOutboxTransaction) => Promise<RT>,
+    options: ReadWriteTransactionOptions,
+    runFn: TransactionFn<SpannerOutboxTransaction, RT>,
   ): Promise<RT> {
     return await this.entityManager.transaction(async (dbTransaction) => {
       const stateTransaction = new SpannerStateTransaction(
@@ -52,9 +56,10 @@ export class SpannerOutboxTransactionRunner extends OutboxTransactionRunner<Span
       );
       const eventTransaction = eventTransactionFactory();
 
-      const transaction = new SpannerTransaction(
+      const transaction = new SpannerOutboxTransaction(
         stateTransaction,
         eventTransaction,
+        options.publishOptions,
       );
 
       try {